[IUS-2220] move presenter logic into module
