+++ app/presenters/hyrax/ds_file_set_presenter.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Hyrax
-
   class DsFileSetPresenter < Hyrax::FileSetPresenter
+    include ::Datacore::PresentsArchiveFile
 
     delegate :doi, :doi_the_correct_one,
              :doi_minted?,
@@ -27,40 +27,6 @@ module Hyrax
     # def doi_pending?
     #   @solr_document[ Solrizer.solr_name( 'doi', :symbol ) ].first == ::Deepblue::DoiBehavior::DOI_PENDING
     # end
-
-    # archive files bypass fedora storage
-    def archive_file?
-      mime_type.match(/^message\/external-body\;.*access-type=URL/).present?
-    end
-
-    def archive_request_url
-      return '/' unless archive_file?
-      mime_type.split('"').last
-    end
-
-    def archive_status_url
-      archive_request_url.sub('/request/', '/status/')
-    end
-
-    def archive_file
-      @archive_file ||=
-        if archive_file?
-          collection, object = mime_type.split('"').last.sub('/sda/request/', '').split('/')
-          ArchiveFile.new(collection: collection, object: object)
-        end
-    end
-
-    def archive_status
-      @archive_status ||= archive_file.status
-    end
-
-    def request_action
-      @request_action ||= archive_file.request_action
-    end
-
-    def request_actionable?
-      @request_actionable ||= archive_file.request_actionable?(archive_status)
-    end
 
     def relative_url_root
       rv = ::DeepBlueDocs::Application.config.relative_url_root
@@ -118,7 +84,5 @@ module Hyrax
     def file_size_too_large_to_download?
       !@solr_document.file_size.nil? && @solr_document.file_size >= DeepBlueDocs::Application.config.max_work_file_size_to_download
     end
-
   end
-
 end
